+++ packages/components/src/InputEmail/InputEmail.rebuilt.tsx
@@ -1,13 +1,10 @@
 import React, { forwardRef, useId, useRef } from "react";
 import { useInputEmailActions } from "./hooks/useInputEmailActions";
 import type { InputEmailRebuiltProps } from "./InputEmail.types";
-import {
-  FormFieldWrapper,
-  useAtlantisFormFieldName,
-  useFormFieldWrapperStyles,
-} from "../FormField";
+import { FormFieldWrapper, useAtlantisFormFieldName } from "../FormField";
 import { FormFieldPostFix } from "../FormField/FormFieldPostFix";
 import { filterDataAttributes } from "../sharedHelpers/filterDataAttributes";
+import formFieldStyles from "../FormField/FormField.module.css";
 
 export const InputEmailRebuilt = forwardRef(function InputEmailInternal(
   props: InputEmailRebuiltProps,
@@ -19,18 +16,6 @@ export const InputEmailRebuilt = forwardRef(function InputEmailInternal(
     (ref as React.RefObject<HTMLInputElement>) ??
     useRef<HTMLInputElement>(null);
   const wrapperRef = useRef<HTMLDivElement>(null);
-
-  const { inputStyle } = useFormFieldWrapperStyles({
-    size: props.size,
-    inline: props.inline,
-    align: props.align,
-    type: "email",
-    value: props.value,
-    invalid: props.invalid,
-    error: props.error,
-    disabled: props.disabled,
-    placeholder: props.placeholder,
-  });
 
   const { name } = useAtlantisFormFieldName({
     nameProp: props.name,
@@ -44,6 +29,11 @@ export const InputEmailRebuilt = forwardRef(function InputEmailInternal(
     handleKeyDown,
     handleKeyUp,
     handleClear,
+    handleClick,
+    handleMouseDown,
+    handleMouseUp,
+    handlePointerDown,
+    handlePointerUp,
   } = useInputEmailActions({
     onChange: props.onChange,
     onBlur: props.onBlur,
@@ -51,6 +41,11 @@ export const InputEmailRebuilt = forwardRef(function InputEmailInternal(
     onKeyDown: props.onKeyDown,
     onKeyUp: props.onKeyUp,
     onEnter: props.onEnter,
+    onClick: props.onClick,
+    onMouseDown: props.onMouseDown,
+    onMouseUp: props.onMouseUp,
+    onPointerDown: props.onPointerDown,
+    onPointerUp: props.onPointerUp,
     inputRef,
   });
 
@@ -73,6 +68,7 @@ export const InputEmailRebuilt = forwardRef(function InputEmailInternal(
       description={props.description}
       clearable={props.clearable ?? "never"}
       onClear={handleClear}
+      readonly={props.readOnly}
       wrapperRef={wrapperRef}
       disabled={props.disabled}
       type="email"
@@ -85,7 +81,7 @@ export const InputEmailRebuilt = forwardRef(function InputEmailInternal(
         name={name}
         type="email"
         ref={inputRef}
-        className={inputStyle}
+        className={formFieldStyles.input}
         value={props.value}
         disabled={props.disabled}
         readOnly={props.readOnly}
@@ -110,6 +106,11 @@ export const InputEmailRebuilt = forwardRef(function InputEmailInternal(
         onFocus={handleFocus}
         onKeyDown={handleKeyDown}
         onKeyUp={handleKeyUp}
+        onClick={handleClick}
+        onMouseDown={handleMouseDown}
+        onMouseUp={handleMouseUp}
+        onPointerDown={handlePointerDown}
+        onPointerUp={handlePointerUp}
         data-testid="ATL-InputEmail-input"
         {...dataAttrs}
       />