+++ packages/components/src/InputText/InputText.types.ts
@@ -4,6 +4,7 @@ import type {
   FocusEvents,
   HTMLInputBaseProps,
   KeyboardEvents,
+  MouseEvents,
   RebuiltInputCommonProps,
 } from "../sharedHelpers/types";
 
@@ -18,9 +19,8 @@ export interface RowRange {
  */
 interface InputLengthConstraint {
   /**
-   * Maximum character length for the input.
-   * Note: This affects both HTML validation AND visual width of the component.
-   * Future work will separate these concerns.
+   * The maximum number of characters supported by the input.
+   * @see {@link https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#htmlattrdefmaxlength}
    */
   readonly maxLength?: number;
 }
@@ -33,6 +33,7 @@ export type InputTextVersion = 1 | 2 | undefined;
  */
 export interface InputTextRebuiltProps
   extends HTMLInputBaseProps,
+    MouseEvents<HTMLInputElement | HTMLTextAreaElement>,
     FocusEvents<HTMLInputElement | HTMLTextAreaElement>,
     KeyboardEvents<HTMLInputElement | HTMLTextAreaElement>,
     RebuiltInputCommonProps,