+++ packages/components/src/InputTime/InputTime.rebuilt.test.tsx
@@ -109,8 +109,6 @@ describe("InputTimeRebuilt", () => {
       (TimePredictHook.useTimePredict as jest.Mock).mockClear();
     });
 
-    // This is asserting imperfect behavior that is preexisting
-    // ideally it would only be called once
     it("calls useTimePredict hook on initial render", () => {
       const handleChange = jest.fn();
 
@@ -123,7 +121,7 @@ describe("InputTimeRebuilt", () => {
         />,
       );
 
-      expect(TimePredictHook.useTimePredict).toHaveBeenCalledTimes(2);
+      expect(TimePredictHook.useTimePredict).toHaveBeenCalledTimes(1);
       expect(TimePredictHook.useTimePredict).toHaveBeenCalledWith({
         value: initialValue,
         handleChange: expect.any(Function),
@@ -241,6 +239,27 @@ describe("InputTimeRebuilt", () => {
       await userEvent.type(input, "1");
 
       expect(mockSetTypedTime).not.toHaveBeenCalled();
+    });
+  });
+
+  describe("Mouse event handlers", () => {
+    it("should call all mouse handlers during a click", async () => {
+      const handlers = {
+        onClick: jest.fn(),
+        onMouseDown: jest.fn(),
+        onMouseUp: jest.fn(),
+        onPointerDown: jest.fn(),
+        onPointerUp: jest.fn(),
+      };
+      render(<InputTime version={2} {...handlers} />);
+      const input = screen.getByTestId("ATL-InputTime-input");
+      await userEvent.click(input);
+
+      expect(handlers.onClick).toHaveBeenCalledTimes(1);
+      expect(handlers.onMouseDown).toHaveBeenCalledTimes(1);
+      expect(handlers.onMouseUp).toHaveBeenCalledTimes(1);
+      expect(handlers.onPointerDown).toHaveBeenCalledTimes(1);
+      expect(handlers.onPointerUp).toHaveBeenCalledTimes(1);
     });
   });
 });